Swift 2: Revised error system (due to ErrorType)
Since ErrorType is not (necessarily) an NSObject we needed to handle the unhandled tokens differently.

Instead we keep an ErrorConsumptionToken with the ErrorType and ensure we pass it around with subsequent promises.

This appears to work for all possible Swift paths and thus I think it will work for all ObjC paths too and it may be worth switching to this system.

The reason the other system was used (add a runtime iVar) is so that errors that are consumed but then end up going back into the a chain are known to already be consumed. However with further realization it is apparent that any error that goes back into a chain is unconsumed and thus the new system should be work well in all cases.