Remove ObjC PMKSetUnhandledErrorHandler
If you need this, you must write some Swift. The ObjC simply doesn’t work with the Swift 2 beta due to the implicit ErrorType -> NSError conversions and this seems a good enough opportunity to cut this code. Since the function is one-time use per app it seems the inconvenience for an ObjC developer is acceptable.