Fixes #201
It’s enough to make the associated reference RETAIN.

I used this answer to verify that my accesses to the consumable are now safe: http://stackoverflow.com/a/29935110/6444

Now it is still possible for the consumed property to be set out-of-order or even partially, but protecting this just doesn’t seem worth it. I’m not even sure what *should* happen in a situation where two chains with the same error resolution consume and unconsume the same error. Either order seems like luck of the draw and should it be consumed or not? Hard to say.